Add files via upload
this is a program, which reads the grades of 10 college students for 3 courses, its student
besides of providing the grade, he/she will be asked to provide their last name
as well. The program will calculate the average grade of each student and
will inform us about the maximum average and the minimum.